requirements.txt
torch>=1.8.0
torchvision>=0.9.0
rdkit>=2022.03.4
scikit-learn>=0.24.2
networkx>=2.6.3
pandarallel>=1.5.3
mne>=1.0.3
urllib3<=1.26.15
numpy
pandas
tqdm
pyyaml
ogb
openai